Edit javascript.js for computer choice/response
I did a lot of work and forgot to commit for a while. In any case,
I've tested the code in my console and it seems okay to me. If I were to
screw the code up from here on out, I could always just revert back to
this version.

These are the first edits I ma